fixed issue with mistyped validationErrorTo
If validationError has been mistyped, the code will not find requested
element.
It will then just append a new element as is standard functionality.
But it will never try to find the added element if it revalidates, and
add a new message for each validation. This should fix that problem