Fixed bug #16 reset ValidationSummary route change
- Added option to display only last error message instead of all
messages at once
- Added new template to display ngRepeat examples, also help
troubleshooting the follwoing bug.
- Fixed a bug (probably linked to issue #16) where changing route on
View/Controller would make the ValidationSummary fail when coming back
to original View/Controller, this bug was associated to the fact that
the ValidationSummary kept growing from Controller to Controller, now
this ValidationSummary is wipe out as soon as we detect a route change.