Fixed an issue with 3rd party single object
- Fixed an issue with 3rd validation when an object was disguised as an
array and was not triggering a $scope.$watch, so the validation was
never kicking in. For example the 3rd party package "Dropdown
Multiselect" when setting the option of "selectionLimit" to 1 is
returning an array while in fact it's a an object. It's returning this
`[id: 1, label: 'John']` but in reality this is an object (not an array)
and should be this `{id: 1, label: 'John'}` and so because of that, the
$scope.$watch never kicks in.
- Also added a international phone number validation