Fix checkFormValidity formName #135 #139 #140 #141
CheckFormValidity was not working in all cases, there was a problem
inside the function `getElementParentForm( )` which was returning null
very often. This is due to the fact that `.form` only works with `input`
element and so if user had validation on let say a `<div>` or any other
Angular element that isn't an input then the `getElementParentForm( )`
was returning null which was in turn making the
`checkFormValidity($scope.formName)` and `$validationSummary` not
working correctly.